refactor: rename `websocket-client.ts` to `startTerminal.ts`
Websockets have not been used for a long time, so the file name is
misleading.
