+++ packages/library/src/browser/neovim-client.ts
@@ -1,7 +1,7 @@
 import type { Terminal } from "@xterm/xterm"
 import { TerminalClient as NeovimTerminalClient } from "../client/index.js"
+import type { TuiTerminalApi } from "../client/startTerminal.js"
 import { TerminalTerminalClient } from "../client/terminal-terminal-client.js"
-import type { TuiTerminalApi } from "../client/websocket-client.js"
 import type {
   ExCommandClientInput,
   LuaCodeClientInput,
