+++ packages/library/src/client/startTerminal.ts
